TagV2
So I don't really know how to make a draft of one of these I couldn't get the branches to compare and stuff...
I plan on adding more syntax as I see fit, read the comment at the top for the changes I have made. Indigo is my staple, please leave it as indigo.