Message to Converter
### Short explanation
This is a little command that I made for use in a Discord server. It uses some code from the example commands, but what this does is it identifies the contents of the message, and based on that, it either contains the attachment (image) from the message and the text, or it contains the text only if there are no attachments.
### Trigger
The trigger for this custom command is _Regex_ **.***
### Optimization
This might require a lot of changes or there are better ways to do this, but I couldn't find an example that would convert any message to an embed regardless if it was an image, or text, or text and image.
### Prospects
I believe the future of this could be a command that can be used in a private channel to configure in which channels is it used, what emojis are used, and perhaps even scheduling, if we were to go all-in. 

**Please describe the changes this PR makes and why it should be merged:**

**Status**

- [x] Code changes have been tested on an instance of YAGPDB, or there are no code changes
- [x] I have read and followed the [contribution guide](../CONTRIBUTING.md)