Create transferroles.go.tmpl
This is my 2nd PR here. I worked on a command that makes YAGPDB give one member all roles another member has. (I only used the sleep function in line 38 because it was really buggy without it.) If I should not respond to a comment, please ping me on discord! Thanks for your time.