chore(overview): clarify code snippets licensing
This is totally my fault here - neglected to contact the other writers of code snippets before changing the license. This commit clarifies that the Unlicense should only apply to snippets I (jo3-l) have written. If you've written a code snippet in the past and want it to be under the Unlicense, please contact me.