Update afk.go.tmpl
When using " as prefix it would cause an issue. This patch should fix that. 
It is not really a good patch as it doesn't solve the root problem of it but it still works. Thats why i am not pushing it and just creating a pr for someone else to review.