website: fix embed slug, update source repo URL
The currently linked URL to YAGPDB's source code is outdated and might stop working at
any time. This commit updates the website to link to the "new" repository hosted under
botlabs-gg.

Furthermore, the introduction is modified as such that the embed generated by hotlinking
the website is properly formatted. Before this commit, the embed's description read as
follows:

    A curated collection of over custom commands for the YAGPDB Discord Bot to enhance
    your server. It is:

With this commit, the description now reads as such:

    A curated collection of over 60 custom commands for the YAGPDB Discord Bot to enhance
    your server.