utilities/time: defend against downtimes
Sometimes wttr.in might run out of queries, thus breaking this custom
command. This patch modifies it as such that it looks for the first line
announcing the downtime, and, if found, returns with an error stating
what happened.

Previously, the custom command simply failed, which isn't very nice UX.

Closes #263

Signed-off-by: Luca Zeuch <l-zeuch@email.de>