docs: remove official remark (#310)
* docs: remove official remark

Remove the "official" remark and replace it with the similar, but less
assertive term "vetted".

Add to the "Need help" section that not every staff may know about every
custom command present here.

The rationale behind this change is the fact that "official" implies
that all staff are more or less knowledgeable about the codebase present
in this collection, which is simply not true.
Further, it cannot be reasonably expected that all staff *are* indeed
knowledgeable about this collection.

"Vetted" has a similar meaning, though it is less assertive:
Someone with good knowledge has looked at the code and confirmed it to
be -- to the best of their abilities -- free of malicious code.

Third-party content always has a risk with it, which is also noted down
(though to a different extent) in LICENSE.md. This change should make it
clearer that whilst official staff *looked* at it, it doesn't mean that
we also serve a guarantee for it.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* style(*): run prettier on codebase

---------

Signed-off-by: Luca Zeuch <l-zeuch@email.de>
Co-authored-by: l-zeuch <l-zeuch@users.noreply.github.com>