starboardv2: fix attachments not embedding (#319)
Fix attachments not embedding properly due to no match being found.

Discord recently changed how attachment URLs are constructed[^1]. They
now have URL parameters, which the current implementation does not
account for.
Removing the constraint to match the file extension and then an EOL
reportedly fixes this issue[^2].

I decided against a more complex regex to match against the URL
parameters as well. If this causes undesirable behaviour in the future,
it is obviously subject to change.

[^1]: https://www.bleepingcomputer.com/news/security/discord-will-switch-to-temporary-file-links-to-block-malware-delivery/
[^2]: https://discord.com/channels/166207328570441728/384011387132706816/1170826121462415400

Reported-by: majorleaguebutters (Discord)
Reported-by: @FravBox

Tested-by: majorleaguebutters (Discord)
Tested-by: m0xx_ (Discord)

Thanks-to: @FravBox

Signed-off-by: Luca Zeuch <l-zeuch@email.de>