utils/snipe: update regex to new undelete layout
YAGPDB version 2.44.0 introduced a new layout for `undelete` using
Discord's dynamic timestamps `<t:UNIX:MOD>`, which the RegEx
(for obvious reasons) cannot account for, causing the `snipe` command to
always respond with "Nothing to snipe here".

Adjust the RegEx to account for this new format and make the rest of the
code work well with the additional data provided.

Closes #341

Signed-off-by: Luca Zeuch <l-zeuch@email.de>