fun/coinflip: parsing change continued
No/invalid bet error included the parsed bet before using `index .CmdArgs 1`. As the new parsing method allows for the command arguments to be in any position, we are unable to pin point which argument is meant to be the bet when a valid bet is not provided. For this reason, a general error stating the min and max bets is used.